Renamed secure_document_submission components
