+++ frontend/src/secure_document_upload/SecureDocumentSubmissionAddContainer.js
@@ -3,104 +3,65 @@
  * All data handling & manipulation should be handled here.
  */
 import PropTypes from 'prop-types';
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import { bindActionCreators } from 'redux';
+import React, {Component} from 'react';
+import {connect} from 'react-redux';
+import {bindActionCreators} from 'redux';
 
+import CreditTransactionRequestForm from './components/SecureDocumentSubmissionForm';
 import Loading from '../app/components/Loading';
 import Modal from '../app/components/Modal';
-import CreditTransactionRequestForm from './components/CreditTransactionRequestForm';
-
+import history from '../app/History';
 import {
-  deleteDocumentUpload, getDocumentUpload, getDocumentUploadURL, partialUpdateDocument,
+  addDocumentUpload,
+  clearDocumentUploadError,
+  getDocumentUploadURL,
   uploadDocument
 } from '../actions/documentUploads';
-import history from '../app/History';
 import DOCUMENT_STATUSES from '../constants/documentStatuses';
 import SECURE_DOCUMENT_UPLOAD from '../constants/routes/SecureDocumentUpload';
 import toastr from '../utils/toastr';
+import FileUploadProgress from "./components/FileUploadProgress";
 
-class CreditTransactionRequestEditContainer extends Component {
-  constructor (props) {
+class SecureDocumentSubmissionAddContainer extends Component {
+  constructor(props) {
     super(props);
 
     this.state = {
       fields: {
         attachmentCategory: '',
         attachments: [],
-        compliancePeriod: { id: 0, description: '' },
+        comment: '',
+        compliancePeriod: {id: 0, description: ''},
         documentType: {
-          id: 1
+          id: props.match.params.type ? parseInt(props.match.params.type, 10) : 1
         },
         files: [],
         milestone: '',
         title: ''
       },
+      validationErrors: {},
       uploadState: '',
-      validationErrors: {}
+      uploadProgress: []
     };
-
-    this.loaded = false;
-    this.originalAttachments = [];
 
     this._handleInputChange = this._handleInputChange.bind(this);
     this._handleSubmit = this._handleSubmit.bind(this);
   }
 
-  componentDidMount () {
-    this.loadData(this.props.match.params.id);
+  componentDidMount() {
+    this.props.clearDocumentUploadError();
   }
 
-  componentWillReceiveProps (props) {
-    this.loadPropsToFieldState(props);
-  }
+  changeObjectProp(id, name) {
+    const fieldState = {...this.state.fields};
 
-  changeObjectProp (id, name) {
-    const fieldState = { ...this.state.fields };
-
-    fieldState[name] = { id: id || 0 };
+    fieldState[name] = {id: id || 0};
     this.setState({
       fields: fieldState
     });
   }
 
-  loadData (id) {
-    this.props.getDocumentUpload(id);
-  }
-
-  loadPropsToFieldState (props) {
-    const { item } = props;
-
-    if (Object.keys(item).length > 0 && !this.loaded) {
-      const fieldState = {
-        attachments: item.attachments, // updated source to be compared with the original
-        comment: (item.comments.length > 0) ? item.comments[0].comment : '',
-        compliancePeriod: item.compliancePeriod,
-        documentType: item.type,
-        files: [],
-        milestone: (item.milestone ? item.milestone.milestone : '') || '',
-        title: item.title
-      };
-
-      // original source to see if something was removed
-      this.originalAttachments = item.attachments.slice(0);
-
-      this.setState({
-        fields: fieldState
-      });
-
-      this.loaded = true;
-    }
-  }
-
-  _deleteCreditTransferRequest (id) {
-    this.props.deleteDocumentUpload(id).then(() => {
-      history.push(SECURE_DOCUMENT_UPLOAD.LIST);
-      toastr.documentUpload(null, 'Draft deleted.');
-    });
-  }
-
-  _getDocumentType () {
+  _getDocumentType() {
     let documentTypes = [];
     this.props.referenceData.documentCategories.forEach((category) => {
       documentTypes = documentTypes.concat(category.types);
@@ -115,7 +76,7 @@ class CreditTransactionRequestEditContainer extends Component {
     return false;
   }
 
-  _getErrors () {
+  _getErrors() {
     if ('title' in this.props.errors && this._getDocumentType().theType === 'Evidence') {
       this.props.errors.title.forEach((error, index) => {
         this.props.errors.title[index] = error.replace(/Title/, 'Part 3 Agreement');
@@ -125,13 +86,35 @@ class CreditTransactionRequestEditContainer extends Component {
     return this.props.errors;
   }
 
-  _handleInputChange (event) {
-    const { value, name } = event.target;
-    const fieldState = { ...this.state.fields };
+  _handleInputChange(event) {
+    const {value, name} = event.target;
+    const fieldState = {...this.state.fields};
 
     if (typeof fieldState[name] === 'object' &&
-      name !== 'files' && name !== 'attachments') {
+      name !== 'files') {
       this.changeObjectProp(parseInt(value, 10), name);
+    } else if (name === 'files') {
+        const progress = [];
+        fieldState[name] = value;
+
+        for (let i = 0; i < value.length; i++) {
+            progress.push({
+              index: i,
+              started: false,
+              complete: false,
+              error: false,
+              progress: {
+                loaded: 0,
+                total: 1,
+              }
+            });
+          }
+
+          this.setState({
+            ...this.state,
+            uploadProgress: progress,
+            fields: fieldState,
+          });
     } else {
       fieldState[name] = value;
       this.setState({
@@ -140,10 +123,11 @@ class CreditTransactionRequestEditContainer extends Component {
     }
   }
 
-  _handleSubmit (event, status) {
+  _handleSubmit(event, status) {
     event.preventDefault();
 
     this.setState({
+      ...this.state,
       uploadState: 'progress'
     });
 
@@ -151,54 +135,101 @@ class CreditTransactionRequestEditContainer extends Component {
     const attachments = [];
     const attachedFiles = this.state.fields.files;
 
-    Object.keys(attachedFiles).forEach((file) => {
+    for (let i = 0; i < attachedFiles.length; i++) {
+
+      let file = attachedFiles[i];
+
       uploadPromises.push(new Promise((resolve, reject) => {
         const reader = new FileReader();
+
         reader.onload = () => {
           const blob = reader.result;
 
           this.props.requestURL().then((response) => {
-            uploadDocument(response.data.put, blob);
+
+            let uploadProgress = this.state.uploadProgress;
+            uploadProgress[i] = {
+              ...uploadProgress[i],
+              started: true
+            };
+            this.setState({
+              ...this.state,
+              uploadProgress
+            });
+
+
+            this.props.uploadDocument(response.data.put, blob,
+              (progressEvent) => {
+                let uploadProgress = this.state.uploadProgress;
+                uploadProgress[i] = {
+                  ...uploadProgress[i],
+                  progress: {
+                    loaded: progressEvent.loaded,
+                    total: progressEvent.total
+                  }
+                };
+                this.setState({
+                  ...this.state,
+                  uploadProgress
+                });
+              }
+            ).then(
+              () => {
+                let uploadProgress = this.state.uploadProgress;
+                uploadProgress[i] = {
+                  ...uploadProgress[i],
+                  complete: true,
+                  error: false
+                };
+                this.setState({
+                  ...this.state,
+                  uploadProgress
+                });
+              }
+            ).catch(
+              () => {
+                let uploadProgress = this.state.uploadProgress;
+                uploadProgress[i] = {
+                  ...uploadProgress[i],
+                  complete: false,
+                  error: true
+                };
+                this.setState({
+                  ...this.state,
+                  uploadProgress
+                });
+              }
+            );
 
             attachments.push({
-              filename: attachedFiles[file].name,
-              mimeType: attachedFiles[file].type,
-              size: attachedFiles[file].size,
+              filename: file.name,
+              mimeType: file.type,
+              size: file.size,
               url: response.data.get
             });
             resolve();
           });
         };
 
-        reader.readAsArrayBuffer(attachedFiles[file]);
+        reader.readAsArrayBuffer(file);
       }));
-    });
+    }
 
-    // goes through the original attachments and see if something is missing
-    // then we just fetch the id of the missing ones (ie removed), and send that to the backend
-    const attachmentsToBeRemoved = this.originalAttachments
-      .filter(originalAttachment => (
-        this.state.fields.attachments.findIndex(attachment => (
-          attachment.id === originalAttachment.id)) < 0
-      ))
-      .map(attachment => attachment.id);
-
-    const { id } = this.props.item;
 
     // API data structure
     const data = {
-      attachments,
-      attachmentsToBeRemoved,
       comment: this.state.fields.comment,
       compliancePeriod: this.state.fields.compliancePeriod.id,
       milestone: this.state.fields.milestone,
       status: status.id,
-      title: this.state.fields.title
+      title: this.state.fields.title,
+      type: this.state.fields.documentType.id,
+      attachments
     };
 
     Promise.all(uploadPromises).then(() => (
-      this.props.partialUpdateDocument(id, data).then((response) => {
-        this.setState({ uploadState: 'success' });
+      this.props.addDocumentUpload(data).then((response) => {
+        this.setState({uploadState: 'success'});
         history.push(SECURE_DOCUMENT_UPLOAD.LIST);
         toastr.documentUpload(status.id);
       }).catch((reason) => {
@@ -215,18 +246,18 @@ class CreditTransactionRequestEditContainer extends Component {
     return true;
   }
 
-  render () {
-    if (this.state.uploadState === 'progress' || this.props.referenceData.isFetching || !this.loaded) {
-      return (<Loading />);
+  render() {
+    if (this.props.referenceData.isFetching) {
+      return (<Loading/>);
     }
-    const { item } = this.props;
-    let availableActions = [];
 
-    if (item.actions) {
-      availableActions = item.actions.map(action => (
-        action.status
-      ));
+    if (this.state.uploadState === 'progress') {
+      return (<FileUploadProgress
+        progress={this.state.uploadProgress}
+        files={this.state.fields.files}/>);
     }
+
+    const availableActions = ['Draft', 'Submitted'];
 
     return ([
       <CreditTransactionRequestForm
@@ -234,7 +265,6 @@ class CreditTransactionRequestEditContainer extends Component {
         availableActions={availableActions}
         categories={this.props.referenceData.documentCategories}
         documentType={this._getDocumentType()}
-        edit
         errors={this._getErrors()}
         fields={this.state.fields}
         handleInputChange={this._handleInputChange}
@@ -250,33 +280,23 @@ class CreditTransactionRequestEditContainer extends Component {
         id="confirmSubmit"
         key="confirmSubmit"
       >
-        Are you sure you want to update this request?
-      </Modal>,
-      <Modal
-        handleSubmit={() => this._deleteCreditTransferRequest(item.id)}
-        id="confirmDelete"
-        key="confirmDelete"
-      >
-        Are you sure you want to delete this draft?
+        Are you sure you want to securely submit these files to the Government of British Columbia?
       </Modal>
     ]);
   }
 }
 
-CreditTransactionRequestEditContainer.defaultProps = {
+SecureDocumentSubmissionAddContainer.defaultProps = {
   errors: {},
   validationErrors: {}
 };
 
-CreditTransactionRequestEditContainer.propTypes = {
-  deleteDocumentUpload: PropTypes.func.isRequired,
+SecureDocumentSubmissionAddContainer.propTypes = {
+  addDocumentUpload: PropTypes.func.isRequired,
+  clearDocumentUploadError: PropTypes.func.isRequired,
   errors: PropTypes.shape({
     title: PropTypes.arrayOf(PropTypes.string)
   }),
-  getDocumentUpload: PropTypes.func.isRequired,
-  item: PropTypes.shape({
-    id: PropTypes.number
-  }).isRequired,
   loggedInUser: PropTypes.shape({
     displayName: PropTypes.string,
     hasPermission: PropTypes.func,
@@ -288,22 +308,23 @@ CreditTransactionRequestEditContainer.propTypes = {
   }).isRequired,
   match: PropTypes.shape({
     params: PropTypes.shape({
-      id: PropTypes.string.isRequired
-    }).isRequired
+      id: PropTypes.string,
+      type: PropTypes.string
+    })
   }).isRequired,
   referenceData: PropTypes.shape({
     documentCategories: PropTypes.arrayOf(PropTypes.shape),
     isFetching: PropTypes.bool,
     isSuccessful: PropTypes.bool
   }).isRequired,
+  validationErrors: PropTypes.shape(),
   requestURL: PropTypes.func.isRequired,
-  partialUpdateDocument: PropTypes.func.isRequired,
-  validationErrors: PropTypes.shape()
+  uploadDocument: PropTypes.func.isRequired
+
 };
 
 const mapStateToProps = state => ({
   errors: state.rootReducer.documentUpload.errors,
-  item: state.rootReducer.documentUpload.item,
   loggedInUser: state.rootReducer.userRequest.loggedInUser,
   referenceData: {
     documentCategories: state.rootReducer.referenceData.data.documentCategories,
@@ -313,10 +334,10 @@ const mapStateToProps = state => ({
 });
 
 const mapDispatchToProps = dispatch => ({
-  deleteDocumentUpload: bindActionCreators(deleteDocumentUpload, dispatch),
-  getDocumentUpload: bindActionCreators(getDocumentUpload, dispatch),
+  addDocumentUpload: bindActionCreators(addDocumentUpload, dispatch),
+  clearDocumentUploadError: bindActionCreators(clearDocumentUploadError, dispatch),
   requestURL: bindActionCreators(getDocumentUploadURL, dispatch),
-  partialUpdateDocument: bindActionCreators(partialUpdateDocument, dispatch)
+  uploadDocument: bindActionCreators(uploadDocument, dispatch)
 });
 
-export default connect(mapStateToProps, mapDispatchToProps)(CreditTransactionRequestEditContainer);
+export default connect(mapStateToProps, mapDispatchToProps)(SecureDocumentSubmissionAddContainer);
