+++ frontend/src/secure_document_upload/SecureDocumentSubmissionContainer.js
@@ -8,11 +8,12 @@ import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import PropTypes from 'prop-types';
 
-import CreditTransactionRequestsPage from './components/CreditTransactionRequestsPage';
+
 
 import { getDocumentUploads, getDocumentUploadURL } from '../actions/documentUploads';
+import SecureDocumentSubmissionsPage from './components/SecureDocumentSubmissionsPage';
 
-class CreditTransactionRequestsContainer extends Component {
+class SecureDocumentSubmissionContainer extends Component {
   constructor (props) {
     super(props);
     this.state = {
@@ -29,7 +30,7 @@ class CreditTransactionRequestsContainer extends Component {
 
   render () {
     return (
-      <CreditTransactionRequestsPage
+      <SecureDocumentSubmissionsPage
         categories={this.props.referenceData.documentCategories}
         documentUploads={this.props.documentUploads}
         loggedInUser={this.props.loggedInUser}
@@ -40,10 +41,10 @@ class CreditTransactionRequestsContainer extends Component {
   }
 }
 
-CreditTransactionRequestsContainer.defaultProps = {
+SecureDocumentSubmissionContainer.defaultProps = {
 };
 
-CreditTransactionRequestsContainer.propTypes = {
+SecureDocumentSubmissionContainer.propTypes = {
   documentUploads: PropTypes.shape({
     isFetching: PropTypes.bool,
     items: PropTypes.arrayOf(PropTypes.shape())
@@ -76,4 +77,4 @@ const mapDispatchToProps = dispatch => ({
   requestURL: bindActionCreators(getDocumentUploadURL, dispatch)
 });
 
-export default connect(mapStateToProps, mapDispatchToProps)(CreditTransactionRequestsContainer);
+export default connect(mapStateToProps, mapDispatchToProps)(SecureDocumentSubmissionContainer);
