+++ frontend/src/secure_document_upload/components/LinkedCreditTransferSelection.js
@@ -2,11 +2,8 @@ import React, {Component} from 'react';
 import PropTypes from 'prop-types';
 import FontAwesomeIcon from '@fortawesome/react-fontawesome';
 import * as Lang from '../../constants/langEnUs';
-import LocalTimestamp from '../../app/components/LocalTimestamp';
-import CreditTransactionRequestCommentForm from './CreditTransactionRequestCommentForm';
+import SecureDocumentSubmissionCommentForm from './SecureDocumentSubmissionCommentForm';
 import ReactTable from "react-table";
-import SECURE_DOCUMENT_UPLOAD from "../../constants/routes/SecureDocumentUpload";
-import history from "../../app/History";
 import moment from "moment";
 import {Link} from "react-router-dom";
 import {CREDIT_TRANSFER_STATUS, CREDIT_TRANSFER_TYPES} from "../../constants/values";
