+++ frontend/src/secure_document_upload/components/SecureDocumentSubmissionFormDetails.js
@@ -9,7 +9,7 @@ import FontAwesomeIcon from '@fortawesome/react-fontawesome';
 import CONFIG from '../../config';
 import { getFileSize, getIcon, getScanStatusIcon, validateFiles } from '../../utils/functions';
 
-class CreditTransactionRequestFormDetails extends Component {
+class SecureDocumentSubmissionFormDetails extends Component {
   constructor (props) {
     super(props);
 
@@ -349,12 +349,12 @@ class CreditTransactionRequestFormDetails extends Component {
   }
 }
 
-CreditTransactionRequestFormDetails.defaultProps = {
+SecureDocumentSubmissionFormDetails.defaultProps = {
   children: null,
   edit: false
 };
 
-CreditTransactionRequestFormDetails.propTypes = {
+SecureDocumentSubmissionFormDetails.propTypes = {
   children: PropTypes.oneOfType([
     PropTypes.arrayOf(PropTypes.node),
     PropTypes.node
@@ -386,4 +386,4 @@ CreditTransactionRequestFormDetails.propTypes = {
   handleInputChange: PropTypes.func.isRequired
 };
 
-export default CreditTransactionRequestFormDetails;
+export default SecureDocumentSubmissionFormDetails;
