+++ frontend/src/secure_document_upload/components/SecureDocumentSubmissionTable.js
@@ -12,7 +12,7 @@ import moment from 'moment';
 import history from '../../app/History';
 import SECURE_DOCUMENT_UPLOAD from '../../constants/routes/SecureDocumentUpload';
 
-const CreditTransferRequestTable = (props) => {
+const SecureDocumentSubmissionTable = (props) => {
   const columns = [{
     accessor: 'id',
     className: 'col-id',
@@ -136,9 +136,9 @@ const CreditTransferRequestTable = (props) => {
   );
 };
 
-CreditTransferRequestTable.defaultProps = {};
+SecureDocumentSubmissionTable.defaultProps = {};
 
-CreditTransferRequestTable.propTypes = {
+SecureDocumentSubmissionTable.propTypes = {
   items: PropTypes.arrayOf(PropTypes.shape({
     createUser: PropTypes.shape({
       organization: PropTypes.shape({
@@ -160,4 +160,4 @@ CreditTransferRequestTable.propTypes = {
   }).isRequired
 };
 
-export default CreditTransferRequestTable;
+export default SecureDocumentSubmissionTable;
