+++ frontend/src/secure_document_upload/components/SecureDocumentSubmissionsPage.js
@@ -7,9 +7,9 @@ import Loading from '../../app/components/Loading';
 import * as Lang from '../../constants/langEnUs';
 import history from '../../app/History';
 import PERMISSIONS_SECURE_DOCUMENT_UPLOAD from '../../constants/permissions/SecureDocumentUpload';
-import CreditTransferRequestTable from './CreditTransferRequestTable';
+import SecureDocumentSubmissionTable from './SecureDocumentSubmissionTable';
 
-const CreditTransactionRequestsPage = (props) => {
+const SecureDocumentSubmissionsPage = (props) => {
   const { isFetching, items } = props.documentUploads;
   const isEmpty = items.length === 0;
 
@@ -61,7 +61,7 @@ const CreditTransactionRequestsPage = (props) => {
       </div>
       {isFetching && <Loading />}
       {!isFetching &&
-      <CreditTransferRequestTable
+      <SecureDocumentSubmissionTable
         items={items}
         isFetching={isFetching}
         isEmpty={isEmpty}
@@ -72,10 +72,10 @@ const CreditTransactionRequestsPage = (props) => {
   );
 };
 
-CreditTransactionRequestsPage.defaultProps = {
+SecureDocumentSubmissionsPage.defaultProps = {
 };
 
-CreditTransactionRequestsPage.propTypes = {
+SecureDocumentSubmissionsPage.propTypes = {
   categories: PropTypes.arrayOf(PropTypes.shape()).isRequired,
   documentUploads: PropTypes.shape({
     isFetching: PropTypes.bool,
@@ -87,4 +87,4 @@ CreditTransactionRequestsPage.propTypes = {
   title: PropTypes.string.isRequired
 };
 
-export default CreditTransactionRequestsPage;
+export default SecureDocumentSubmissionsPage;
